itest: put mpp tests in one file
